+++ AppLinking/Editor/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Runtime.CompilerServices;
 using Unity.Cloud.Common;
 
-[assembly: ApiSourceVersion("com.unity.cloud.app-linking.editor", "1.1.5")]
+[assembly: ApiSourceVersion("com.unity.cloud.app-linking.editor", "1.2.0-exp.3")]
 [assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Editor")]
 [assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Tests.Editor")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]